make it responsive in mobile

+++ lib/features/app/pages/manage_hostel/presentation/widgets/custom_my_hostel_card.dart
@@ -21,7 +21,7 @@ class CustomMyHostelCard extends StatelessWidget {
   final String imageUrl;
   final String title;
   final String location;
-  final int rent;
+  final double rent;
   final double rating;
   final int distance;
   final Status status;
